SIMBIOHUB-158: Fix tests not compiling
