+++ api/src/models/survey-create.test.ts
@@ -3,6 +3,7 @@ import { describe } from 'mocha';
 import {
   PostAgreementsData,
   PostLocationData,
+  PostPartnershipsData,
   PostPermitData,
   PostProprietorData,
   PostPurposeAndMethodologyData,
@@ -45,6 +46,10 @@ describe('PostSurveyObject', () => {
 
     it('sets agreements', () => {
       expect(data.agreements).to.equal(null);
+    });
+
+    it('sets partnerships', function () {
+      expect(data.partnerships).to.equal(null);
     });
   });
 
@@ -91,6 +96,45 @@ describe('PostSurveyObject', () => {
 
     it('sets agreements', () => {
       expect(data.agreements).to.instanceOf(PostAgreementsData);
+    });
+  });
+});
+
+describe('PostPartnershipsData', () => {
+  describe('No values provided', () => {
+    let projectPartnershipsData: PostPartnershipsData;
+
+    before(() => {
+      projectPartnershipsData = new PostPartnershipsData(null);
+    });
+
+    it('sets indigenous_partnerships', function () {
+      expect(projectPartnershipsData.indigenous_partnerships).to.eql([]);
+    });
+
+    it('sets stakeholder_partnerships', function () {
+      expect(projectPartnershipsData.stakeholder_partnerships).to.eql([]);
+    });
+  });
+
+  describe('All values provided', () => {
+    let projectPartnershipsData: PostPartnershipsData;
+
+    const obj = {
+      indigenous_partnerships: [1, 2],
+      stakeholder_partnerships: ['partner1, partner2']
+    };
+
+    before(() => {
+      projectPartnershipsData = new PostPartnershipsData(obj);
+    });
+
+    it('sets indigenous_partnerships', function () {
+      expect(projectPartnershipsData.indigenous_partnerships).to.eql(obj.indigenous_partnerships);
+    });
+
+    it('sets stakeholder_partnerships', function () {
+      expect(projectPartnershipsData.stakeholder_partnerships).to.eql(obj.stakeholder_partnerships);
     });
   });
 });
