+++ api/src/models/survey-update.test.ts
@@ -1,6 +1,7 @@
 import { expect } from 'chai';
 import { describe } from 'mocha';
 import {
+  PutPartnershipsData,
   PutSurveyDetailsData,
   PutSurveyLocationData,
   PutSurveyObject,
@@ -474,6 +475,45 @@ describe('PutLocationData', () => {
 
     it('sets revision_count', () => {
       expect(data.revision_count).to.equal(obj.revision_count);
+    });
+  });
+});
+
+describe('PutPartnershipsData', () => {
+  describe('No values provided', () => {
+    let data: PutPartnershipsData;
+
+    before(() => {
+      data = new PutPartnershipsData(null);
+    });
+
+    it('sets indigenous_partnerships', () => {
+      expect(data.indigenous_partnerships).to.eql([]);
+    });
+
+    it('sets stakeholder_partnerships', () => {
+      expect(data.stakeholder_partnerships).to.eql([]);
+    });
+  });
+
+  describe('all values provided', () => {
+    const obj = {
+      indigenous_partnerships: [1, 2],
+      stakeholder_partnerships: ['partner 3', 'partner 4']
+    };
+
+    let data: PutPartnershipsData;
+
+    before(() => {
+      data = new PutPartnershipsData(obj);
+    });
+
+    it('sets indigenous_partnerships', () => {
+      expect(data.indigenous_partnerships).to.eql(obj.indigenous_partnerships);
+    });
+
+    it('sets stakeholder_partnerships', () => {
+      expect(data.stakeholder_partnerships).to.eql(obj.stakeholder_partnerships);
     });
   });
 });
