+++ app/src/features/surveys/view/components/SurveyPartnershipsForm.test.tsx
@@ -1,11 +1,7 @@
 import { IMultiAutocompleteFieldOption } from 'components/fields/MultiAutocompleteFieldVariableSize';
 import { Formik } from 'formik';
 import { render } from 'test-helpers/test-utils';
-import ProjectPartnershipsForm, {
-  IProjectPartnershipsForm,
-  ProjectPartnershipsFormInitialValues,
-  ProjectPartnershipsFormYupSchema
-} from './SurveyPartnershipsForm';
+import SurveyPartnershipsForm, { ISurveyPartnershipsForm, SurveyPartnershipsFormInitialValues, SurveyPartnershipsFormYupSchema } from './SurveyPartnershipsForm';
 
 const first_nations: IMultiAutocompleteFieldOption[] = [
   {
@@ -31,15 +27,37 @@ const stakeholder_partnerships: IMultiAutocompleteFieldOption[] = [
 
 describe('ProjectPartnershipsForm', () => {
   it('renders correctly with default empty values', () => {
+    /*
+    // TODO 
+
+  const codesContext = useContext(CodesContext);
+
+  const codes = codesContext.codesDataLoader.data;
+  codesContext.codesDataLoader.load();
+
+  const { handleSubmit } = formikProps;
+
+  const first_nations: IMultiAutocompleteFieldOption[] =
+    codes?.first_nations?.map((item) => {
+      return { value: item.id, label: item.name };
+    }) || [];
+
+  const stakeholder_partnerships: IMultiAutocompleteFieldOption[] =
+    codes?.agency?.map((item) => {
+      return { value: item.name, label: item.name };
+    }) || [];
+    */
+
+
     const { getByLabelText } = render(
       <Formik
-        initialValues={ProjectPartnershipsFormInitialValues}
-        validationSchema={ProjectPartnershipsFormYupSchema}
+        initialValues={SurveyPartnershipsFormInitialValues}
+        validationSchema={SurveyPartnershipsFormInitialValues}
         validateOnBlur={true}
         validateOnChange={false}
         onSubmit={async () => {}}>
         {() => (
-          <ProjectPartnershipsForm first_nations={first_nations} stakeholder_partnerships={stakeholder_partnerships} />
+          <SurveyPartnershipsForm />
         )}
       </Formik>
     );
@@ -49,7 +67,7 @@ describe('ProjectPartnershipsForm', () => {
   });
 
   it('renders correctly with existing funding values', () => {
-    const existingFormValues: IProjectPartnershipsForm = {
+    const existingFormValues: ISurveyPartnershipsForm = {
       partnerships: {
         indigenous_partnerships: [1, 2],
         stakeholder_partnerships: [1 as unknown as string]
@@ -59,12 +77,12 @@ describe('ProjectPartnershipsForm', () => {
     const { getByLabelText, getByText } = render(
       <Formik
         initialValues={existingFormValues}
-        validationSchema={ProjectPartnershipsFormYupSchema}
+        validationSchema={SurveyPartnershipsFormYupSchema}
         validateOnBlur={true}
         validateOnChange={false}
         onSubmit={async () => {}}>
         {() => (
-          <ProjectPartnershipsForm first_nations={first_nations} stakeholder_partnerships={stakeholder_partnerships} />
+          <SurveyPartnershipsForm />
         )}
       </Formik>
     );
